Fix bugs: UB/data-race/OOM/memory-leak + bounds checks + English comments
- Vec3D.h: remove dead -0.0 filter (no-op under IEEE 754)
- JavaRandom.h: cast to int64_t to fix signed-overflow UB in nextInt
- Block.cpp: guard clause order (y >= 0 before canFallBelow)
- Packet.h: fix writeUTF limit (65535 -> 32767)
- NetHandler.h: unexpectedPacket now kicks instead of silent ignore
- EntityMobs.h: protect lava fire from being cleared by daylight logic
- ItemInWorldManager.cpp: null stack before destroyCurrentEquippedItem (use-after-free)
- AllPackets.h: bounds checks on Packet5/51/52/60 sizes (OOM protection)
- NetLoginHandler.cpp: fix data race on serverId_ (mutex-protected)
- NetLoginHandler.cpp: use unique_ptr for NetServerHandler (leak fix)
- NetworkListenThread.h/cpp: activeConnections_ now unique_ptr (leak fix)
- NetServerHandler.cpp: destructor deletes player_ (EntityPlayerMP leak fix)
- World.cpp: limit spawn search to 10000 iterations (infinite loop fix)
- World.cpp: translate Russian comment to English

+++ src/core/JavaRandom.h
@@ -53,7 +53,7 @@ public:
         do {
             bits = next(31);
             val = bits % bound;
-        } while (bits - val + (bound - 1) < 0);
+        } while (static_cast<int64_t>(bits) - val + (bound - 1) < 0);
         return val;
     }
